bux fixes, improvements and doc

+++ app/src/main/java/com/inovatech/smartpack/model/Service.kt
@@ -1,10 +1,13 @@
 package com.inovatech.smartpack.model
 
 
+/**
+ * Classe que defineix l'entitat Servei
+ */
 data class Service(
     val id: Long = 0,
     val status: ServiceStatus = ServiceStatus.ORDENAT,
-    val userId: Int = 0,
+    val userId: Long = 0,
     val deliverymanId: Long = 0,
     val packageToDeliver: Package = Package()
 )
