+++ app/src/main/java/com/inovatech/smartpack/model/api/ServiceDTO.kt
@@ -4,10 +4,13 @@ import com.google.gson.annotations.SerializedName
 import com.inovatech.smartpack.model.Service
 import com.inovatech.smartpack.model.ServiceStatus
 
+/**
+ * Classe que representa una petició i una resposta per realitzar accions sobre l'entitat Servei
+ */
 data class ServiceDTO(
     @SerializedName("id") val id: Long? = null,
     @SerializedName("estat") val status: ServiceStatus? = null,
-    @SerializedName("usuariId") val userId: Int? = null,
+    @SerializedName("usuariId") val userId: Long? = null,
     @SerializedName("transportistaId") val deliverymanId: Long? = null,
     @SerializedName("paquet") val packageRequest: PackageDTO? = PackageDTO()
 )
